+++ src/extractors.js
@@ -1,6 +1,6 @@
-const safeEval = require('safe-eval');
 const flattenDeep = require('lodash/flattenDeep');
-const tools = require('./tools')
+const vm = require('vm');
+const tools = require('./tools');
 
 // Fetch all main category paths from homepage
 const getAllMainCategoryPaths = ($) => {
@@ -8,14 +8,14 @@ const getAllMainCategoryPaths = ($) => {
 };
 
 // Fetch every subcategory hidden pages (loaders)
-const getAllSubCategories = async ($) => {
-    const dataScript = $($('script').filter((i, script) => $(script).html().includes('runParams')).get()[0]).html();
+const getAllSubCategories = ($) => {
+    const dataScript = $('script').filter((_, script) => $(script).html().includes('runParams')).map((_, el) => $(el).html()).get()[0];
 
     const data = flattenDeep(JSON.parse(
         dataScript.split('window.runParams = ')[2].split('window.runParams.csrfToken =')[0].replace(/;/g, ''),
     ).refineCategory
-        .map(category => category.childCategories))
-        .filter(el => el).map(item => ({ name: item.categoryName, link: `https:${item.categoryUrl}` }));
+        .map((category) => category.childCategories))
+        .filter((el) => el).map((item) => ({ name: item.categoryName, link: `https:${item.categoryUrl}` }));
 
     return data;
 };
@@ -31,7 +31,8 @@ const filterSubCategories = (categoryStartIndex = 0, categoryEndIndex = null, su
 
 // Fetch all products from a global object `runParams`
 const getProductsOfPage = ($) => {
-    const dataScript = $($('script').filter((i, script) => $(script).html().includes('runParams')).get()[0]).html();
+    const dataScript = $('script').filter((_, script) => $(script).html().includes('runParams')).map((_, el) => $(el).html()).get()[0];
+
     const data = JSON.parse(
         dataScript.split('window.runParams = ')[2].split('window.runParams.csrfToken =')[0].replace(/;/g, ''),
     );
@@ -40,14 +41,16 @@ const getProductsOfPage = ($) => {
         throw new Error('We got blocked when trying to fetch products!');
     }
 
-    return data.items && data.items.length > 0 ? data.items.map(item => ({ id: item.productId, name: item.title, link: item.productDetailUrl })) : [];
+    return data.items && data.items.length > 0
+        ? data.items.map((item) => ({ id: item.productId, name: item.title, link: item.productDetailUrl }))
+        : [];
 };
 
 // Fetch basic product detail from a global object `runParams`
 const getProductDetail = async ($, url, extendOutputFunction) => {
-    const dataScript = $($('script').filter((i, script) => $(script).html().includes('runParams')).get()[0]).html();
+    const dataScript = $('script').filter((_, script) => $(script).html().includes('runParams')).map((_, el) => $(el).html()).get()[0];
 
-    const { data } = safeEval(dataScript.split('window.runParams = ')[1].split('var GaData')[0].replace(/;/g, ''));
+    const { data } = vm.compileFunction(`return (${dataScript.split('window.runParams = ')[1].split('var GaData')[0].replace(/;/g, '')})`)();
 
     const {
         actionModule,
@@ -66,7 +69,6 @@ const getProductDetail = async ($, url, extendOutputFunction) => {
         extendOutputFunction = tools.evalExtendOutputFunction(extendOutputFunction);
         extendOutputData = await extendOutputFunction($);
     }
-
 
     return {
         id: actionModule.productId,
@@ -92,33 +94,32 @@ const getProductDetail = async ($, url, extendOutputFunction) => {
             return obj;
         }) : [],
         categories: crossLinkModule.breadCrumbPathList
-            .map(breadcrumb => breadcrumb.target)
-            .filter(breadcrumb => breadcrumb),
+            .map((breadcrumb) => breadcrumb.target)
+            .filter((breadcrumb) => breadcrumb),
         wishedCount: actionModule.itemWishedCount,
         quantity: actionModule.totalAvailQuantity,
         photos: imageModule.imagePathList,
         skuOptions: skuModule.productSKUPropertyList ? skuModule.productSKUPropertyList
-            .map(skuOption => ({
+            .map((skuOption) => ({
                 name: skuOption.skuPropertyName,
                 values: skuOption.skuPropertyValues
-                    .map(skuPropVal => skuPropVal.propertyValueDefinitionName),
+                    .map((skuPropVal) => skuPropVal.propertyValueDefinitionName),
             })) : [],
-        prices: skuModule.skuPriceList.map(skuPriceItem => ({
+        prices: skuModule.skuPriceList.map((skuPriceItem) => ({
             price: skuPriceItem.skuVal.skuAmount.formatedAmount,
             attributes: skuPriceItem.skuPropIds.split(',')
-                .map(propId => {
+                .map((propId) => {
                     const propVal = skuModule.productSKUPropertyList ? skuModule.productSKUPropertyList
-                    .reduce((arr, obj) => { return arr.concat(obj.skuPropertyValues); }, [])
-                    .find(propVal => propVal.propertyValueId === parseInt(propId, 10)) : null
+                        .reduce((arr, obj) => { return arr.concat(obj.skuPropertyValues); }, [])
+                        .find((pVal) => pVal.propertyValueId === +propId) : null;
                     return propVal ? propVal.propertyValueName : null;
                 }),
         })),
         companyId: recommendModule.companyId,
         memberId: commonModule.sellerAdminSeq,
-        ...extendOutputData
+        ...extendOutputData,
     };
 };
-
 
 // Get description HTML of product
 const getProductDescription = async ($) => {
@@ -130,18 +131,18 @@ const getProductFeedback = async ($) => {
     const scrapedFeeds = [];
     for (const item of feedbackItems) {
         const $item = $(item);
-        let infos = $item.find('.user-order-info span').toArray();
+        const infos = $item.find('.user-order-info span').toArray();
         const info = [];
         for (const i of infos) {
             const text = $(i).text().trim().replace(/[\t|\n]+/, '');
             const arr = text.split(':');
             info[arr[0]] = arr[1].trim();
         }
-        let star = undefined;
+        let star;
         const starsWidth = $('.f-rate-info .star-view > span').eq(0).attr('style');
         const pxMatch = starsWidth.match(/\d+/);
         if (pxMatch) {
-            star = 6 - (100 / parseInt(pxMatch[0]));
+            star = 6 - (100 / +pxMatch[0]);
         }
         scrapedFeeds.push({
             userName: $item.find('.user-name').text().trim(),
@@ -149,8 +150,8 @@ const getProductFeedback = async ($) => {
             userStar: star,
             reviewContent: $item.find('.buyer-feedback span:first-child').text().trim(),
             reviewTime: $item.find('.buyer-feedback .r-time-new').text().trim(),
-            info
-        })
+            info,
+        });
     }
     return scrapedFeeds;
 };
@@ -165,20 +166,20 @@ const getProductQA = async (jsonBody) => {
             totalAnswer: question.totalAnswer,
             originalContent: question.content,
             translateContent: question.translateContent,
-            answers: []
-        }
+            answers: [],
+        };
         for (const answers of question.answers) {
             parsedQuestion.answers.push({
                 lang: answers.contentLang,
                 totalAnswer: answers.totalAnswer,
                 originalContent: answers.content,
                 translateContent: answers.translateContent,
-            })
+            });
         }
         productQA.push(parsedQuestion);
     }
     return productQA;
-}
+};
 
 const getTotalQuestions = async (jsonBody) => {
     let totalQuestions = 0;
@@ -187,18 +188,16 @@ const getTotalQuestions = async (jsonBody) => {
         break;
     }
     return totalQuestions;
-}
+};
 
 const getMaxReviews = async ($) => {
     const maxReviewsText = $('.customer-reviews').text();
     const match = maxReviewsText.match(/\d+/g);
-    if (match){
+    if (match) {
         return match[0];
     }
     return 0;
-}
-
-
+};
 
 module.exports = {
     getAllMainCategoryPaths,
@@ -210,5 +209,5 @@ module.exports = {
     getProductFeedback,
     getMaxReviews,
     getProductQA,
-    getTotalQuestions
+    getTotalQuestions,
 };